requirements.txt
flask
dash
pandas
pydo
plotly
python-dotenv